requirements.txt
pandas==1.3.5
requests==2.28.1
json5==0.9.9
numpy==1.21.6
boto3==1.24.51
psycopg2==2.9.3
SQLAlchemy==1.4.40
plotly-express==0.4.1
jupyterlab==3.4.5
dash==2.6.1
sklearn